First commit of effect categories. You can turn it off with EFFECT_CATEGORIES in Experimental.h. Only tested on Linux, there are 4 new files that need to be added to the project files on Win and Mac: src/effects/EffectCategory.[h|cpp] and src/effects/EffectManager.[h|cpp]

+++ sf-cvs/trunk/audacity-src/src/import/ImportPCM.cpp
@@ -256,8 +256,8 @@ int PCMImportFileHandle::Import(TrackFactory *trackFactory,
             framescompleted += block;
          }
 
-         cancelled = !mProgress->Update(framescompleted,
-                                        fileTotalFrames);
+         cancelled = !mProgress->Update((long long unsigned)framescompleted,
+                                        (long long unsigned)fileTotalFrames);
          if (cancelled)
             break;
